Analyze SYSCALL invocations
Added analysis for actual SYSCALL invocations, which now makes
go2seccomp take into account the syscalls made in go's runtime package.

Updated the README with the new output for the examples and talk about
this new analysis and change in limitations.

+++ syscalls.go
@@ -1,7 +1,7 @@
 package main
 
 // reference: https://github.com/seccomp/libseccomp/blob/master/src/arch-x86_64-syscalls.c
-var syscallIDtoName = map[int]string{
+var syscallIDtoName = map[int64]string{
 	0:   "read",
 	1:   "write",
 	2:   "open",
